+++ consensus/consensus_peer.go
@@ -15,12 +15,13 @@ import (
 
 // Consensus Topology Peer
 type ConsensusPeer struct {
+	name    string
 	netAddr types.NetAddress
 	logger  log15.Logger
 	magic   [4]byte
 }
 
-func newConsensusPeer(ip net.IP, port uint16, magic [4]byte) *ConsensusPeer {
+func newConsensusPeer(name string, ip net.IP, port uint16, magic [4]byte) *ConsensusPeer {
 	return &ConsensusPeer{
 		netAddr: types.NetAddress{
 			IP:   ip,
@@ -56,7 +57,7 @@ func (peer *ConsensusPeer) sendData(srcNetAddr types.NetAddress, typeName string
 	// peer.logger.Debug("Send", "data", string(jsonStr), "to", url)
 	_, err = netClient.Post(url, "application/json", bytes.NewBuffer(jsonStr))
 	if err != nil {
-		peer.logger.Error("Failed to send message to peer", "peer", peer.String(), "err", err)
+		peer.logger.Error("Failed to send message to peer", "peer", peer.name, "ip", peer.String(), "err", err)
 		return err
 	}
 	// TODO: check response to verify this action