+++ script/staking/delegates.go
@@ -33,8 +33,8 @@ func newDelegateList(delegates []*Delegate) *DelegateList {
 
 func (d *Delegate) ToString() string {
 	pubKeyEncoded := b64.StdEncoding.EncodeToString(d.PubKey)
-	return fmt.Sprintf("Delegate(Addr=%v, PubKey=%v, Node=%v:%v, VotingPower=%.2e)",
-		d.Address, pubKeyEncoded, string(d.IPAddr), d.Port, float64(d.VotingPower.Int64()))
+	return fmt.Sprintf("Delegate(Name=%v, Addr=%v, PubKey=%v, IP=%v:%v, VotingPower=%d)",
+		string(d.Name), d.Address, pubKeyEncoded, string(d.IPAddr), d.Port, d.VotingPower.String())
 }
 
 func (l *DelegateList) CleanAll() error {
@@ -72,7 +72,7 @@ func (l *DelegateList) ToString() string {
 func GetLatestDelegateList() (*DelegateList, error) {
 	staking := GetStakingGlobInst()
 	if staking == nil {
-		fmt.Println("staking is not initilized...")
+		log.Warn("staking is not initilized...")
 		err := errors.New("staking is not initilized...")
 		return nil, err
 	}
@@ -84,7 +84,7 @@ func GetLatestDelegateList() (*DelegateList, error) {
 	}
 
 	list := staking.GetDelegateList(state)
-	fmt.Println("delegateList from state", list.ToString())
+	// fmt.Println("delegateList from state", list.ToString())
 
 	return list, nil
 }
@@ -106,7 +106,7 @@ func GetInternalDelegateList() ([]*types.Delegate, error) {
 	}
 
 	list := staking.GetDelegateList(state)
-	fmt.Println("delegateList from state", list.ToString())
+	fmt.Println("delegateList from state\n", list.ToString())
 	for _, s := range list.delegates {
 		pubKeyBytes, err := b64.StdEncoding.DecodeString(string(s.PubKey))
 		pubKey, err := crypto.UnmarshalPubkey(pubKeyBytes)