+++ script/staking/stakeholder.go
@@ -28,7 +28,7 @@ func NewStakeholder(holder meter.Address) *Stakeholder {
 func GetLatestStakeholderList() (*StakeholderList, error) {
 	staking := GetStakingGlobInst()
 	if staking == nil {
-		fmt.Println("staking is not initilized...")
+		log.Warn("staking is not initilized...")
 		err := errors.New("staking is not initilized...")
 		return newStakeholderList(nil), err
 	}